refactor(memory): split runner and sandbox modules
