+++ packages/omo-senpi/src/components/memory/worker/runner.ts
@@ -1,47 +1,29 @@
-import { existsSync } from "node:fs"
 import { join } from "node:path"
 
 import type { OmoConfig } from "@oh-my-opencode/omo-config-core"
-import {
-  GitMemoryRepo,
-  buildDefaultSeedFiles,
-  cleanupReflectionWorktree,
-  installHooks,
-  type ReflectionFinalizeResult,
-  type ReflectionWorktree,
-  type ReservedRun,
-} from "@oh-my-opencode/memory-core"
+import type { ReservedRun } from "@oh-my-opencode/memory-core"
 import { loadSenpiOmoConfig, type SenpiOmoConfigResult } from "../../config-resolution"
 import { resolveAgentReflectionSettings } from "../reflection-settings"
 import { createOncePerSessionGuard } from "../../task/usage-guidance"
 import {
   REFLECTION_LAUNCHED_ENTRY_TYPE,
-  recordReflectionCompletion,
   registerReflectionCompletionRenderer,
   safeNotify,
-  type ReflectionCompletionRecord,
   type ReflectionLiveSession,
 } from "./completion"
-import { readReflectionHealth } from "./health"
-import { emitReflectionHealthAlert } from "./health-alert"
 import {
   resolveReflectionModel,
   shouldWarnCategoryUnavailable,
   type ReflectionModelResolution,
 } from "./resolve-model"
-import {
-  MemoryModelExhaustedError,
-  type MemoryModelChain,
-} from "./memory-model-attempts"
-import { resolveAndPreflightMemoryLaunch } from "./memory-launch-preflight"
-import { createRunWorktree } from "./create-run-worktree"
-import { prepareReflectionCandidateSpawn } from "./reflection-spawn-input"
 import { readRunJson } from "./run-artifacts"
-import { failReservationRun, finalizeRecordedOutcome, overrideFailedReservationRun } from "./run-finalization"
 import type { RunFinalizationContext } from "./run-finalization-types"
 import { parseReservationRunLedger } from "./reservation-run-ledger"
-import { requireFinalizedResult } from "./runner-finalization-result"
-import { runReflectionChild } from "./spawn"
+import {
+  publishFinalizedReflectionRun,
+  settleReflectionRun,
+} from "./runner-completion-publication"
+import { executeReflectionRun } from "./runner-execution"
 
 export type {
   ReflectionReservationPort,
@@ -50,7 +32,6 @@ export type {
   SenpiSubprocessRunnerOptions,
 } from "./runner-types"
 
-import { cleanupSucceeded, errorMessage } from "./runner-results"
 import type { ExecutionResult, ReflectionRunResult, ReflectionRunner, SenpiSubprocessRunnerOptions } from "./runner-types"
 
 export class SenpiSubprocessRunner implements ReflectionRunner {
@@ -83,178 +64,47 @@ export class SenpiSubprocessRunner implements ReflectionRunner {
       }, startedAt, resolution, true)
     }
 
-    const execution = await this.execute(run, resolution, loaded, startedAt)
+    const execution = await executeReflectionRun({
+      run,
+      resolution,
+      loaded,
+      startedAt,
+      options: this.options,
+      finalizationContext: () => this.finalizationContext(),
+      appendLaunched: () => this.appendLaunched(run, resolution, startedAt),
+    })
     if ("runId" in execution) return this.deliverFinalized(execution)
-    const settled = await this.settle(run, execution, startedAt, resolution, false)
-    return settled
+    return this.settle(run, execution, startedAt, resolution, false)
   }
 
-  private async execute(
-    run: ReservedRun,
-    resolution: Extract<ReflectionModelResolution, { readonly kind: "resolved" }>,
-    loaded: SenpiOmoConfigResult,
-    startedAt: string,
-  ): Promise<ExecutionResult | ReflectionRunResult> {
-    const repo = new GitMemoryRepo({ dir: this.options.identity.paths.repo, agentId: this.options.identity.id })
-    if (!existsSync(join(this.options.identity.paths.repo, ".git"))) {
-      await repo.init({ seedFiles: buildDefaultSeedFiles(), installHooks: (dir) => { installHooks(dir) } })
-    }
-    let worktree: ReflectionWorktree | undefined
-    try {
-      worktree = await createRunWorktree(repo, run.runId, this.options.identity.paths)
-      const activeWorktree = worktree
-      const reflection = resolveAgentReflectionSettings(loaded.config.memory, this.options.identity.id)
-      const merge = reflection.merge
-      const hardDeadlineAt = Date.now() + (this.options.deadlineMs ?? reflection.timeout_minutes * 60_000)
-      const candidates: MemoryModelChain = [
-        {
-          model: resolution.model,
-          ...(resolution.thinking === undefined ? {} : { thinking: resolution.thinking }),
-        },
-        ...resolution.fallbacks,
-      ]
-      const env = this.options.env ?? process.env
-      let launched = false
-      await (this.options.resolveAndPreflightLaunch ?? resolveAndPreflightMemoryLaunch)({
-        candidates,
-        senpiCommand: this.options.senpiCommand,
-        senpiPrefixArgs: this.options.senpiPrefixArgs,
-        env,
-        envFlag: "SENPI_MEMORY_REFLECTION",
-        configSources: loaded.sources,
-        warn: (message, details) => this.options.logger?.warn(message, details),
-        surfaceName: "reflection",
-        attempt: async (candidate, attemptNumber, nextAttempt) => {
-          const spawnArgs = await prepareReflectionCandidateSpawn({
-            run,
-            worktree: activeWorktree,
-            mergePolicy: merge,
-            category: reflection.category,
-            candidate,
-            attempt: attemptNumber,
-            hardDeadlineAt,
-            nextAttempt,
-            config: loaded.config,
-            identity: this.options.identity,
-            env,
-            senpiCommand: this.options.senpiCommand,
-          })
-          if (!launched) {
-            await this.appendLaunched(run, resolution, startedAt)
-            launched = true
-          }
-          return runReflectionChild(spawnArgs, {
-            terminationGraceMs: this.options.terminationGraceMs,
-            maxOutputBytes: this.options.maxOutputBytes,
-            sandbox: this.options.sandbox,
-            supervisorPath: this.options.supervisorPath,
-          })
-        },
-      })
-    } catch (error) {
-      const runDir = join(this.options.identity.paths.reflection, "runs", run.runId)
-      if (existsSync(join(runDir, "ledger.json"))) {
-        const ledger = parseReservationRunLedger(await readRunJson<unknown>(join(runDir, "ledger.json")))
-        const finalized = error instanceof MemoryModelExhaustedError
-          ? await overrideFailedReservationRun(this.finalizationContext(), runDir, ledger, error.message)
-          : await failReservationRun(
-              this.finalizationContext(),
-              runDir,
-              ledger,
-              "failed",
-              errorMessage(error),
-            )
-        return requireFinalizedResult(finalized)
-      }
-      const discarded = worktree === undefined ? undefined : await this.discard(worktree)
-      return {
-        outcome: "failed",
-        reason: discarded !== undefined && !cleanupSucceeded(discarded) ? "cleanup_failed" : "spawn_failed",
-        detail: [errorMessage(error), discarded?.detail].filter(Boolean).join("; "),
-      }
-    }
-    const runDir = join(this.options.identity.paths.reflection, "runs", run.runId)
-    const ledger = parseReservationRunLedger(await readRunJson<unknown>(join(runDir, "ledger.json")))
-    const finalized = await finalizeRecordedOutcome(this.finalizationContext(), runDir, ledger)
-    return requireFinalizedResult(finalized)
-  }
-
-  private async settle(
+  private settle(
     run: ReservedRun,
     result: ExecutionResult,
     startedAt: string,
     resolution: ReflectionModelResolution,
     suppressCompletionNotification: boolean,
   ): Promise<ReflectionRunResult> {
-    const transition = await this.options.reservation.complete(run.runId, result.outcome)
-    const live = this.options.liveSession?.()
-    this.ensureRenderer(live)
-    const finishedAt = this.now().toISOString()
-    const healthBefore = await readReflectionHealth(join(this.options.identity.paths.reflection, "completions"))
-    const record: ReflectionCompletionRecord = {
-      schemaVersion: 1,
-      runId: run.runId,
-      identity: this.options.identity.id,
-      category: resolution.category,
-      ...(resolution.kind === "resolved"
-        ? {
-            model: result.model ?? resolution.model,
-            ...(result.model === undefined
-              ? resolution.thinking === undefined ? {} : { thinking: resolution.thinking }
-              : result.thinking === undefined ? {} : { thinking: result.thinking }),
-          }
-        : {}),
-      conversationIds: run.request.conversationIds,
-      trigger: run.request.trigger,
-      ...(run.request.trigger === "dream" ? { origin: run.request.origin } : {}),
-      outcome: result.outcome,
-      ...(result.reason === undefined ? {} : { reason: result.reason }),
-      ...(result.detail === undefined ? {} : { detail: result.detail }),
+    return settleReflectionRun({
+      run,
+      result,
       startedAt,
-      finishedAt,
-      durationMs: Math.max(0, Date.parse(finishedAt) - Date.parse(startedAt)),
-      consecutiveFailures: result.outcome === "failed" ? healthBefore.streak + 1 : 0,
-      delivery: { status: "pending" },
-    }
-    const completionsDir = join(this.options.identity.paths.reflection, "completions")
-    const completion = await recordReflectionCompletion(
-      completionsDir,
-      record,
-      suppressCompletionNotification && live
-        ? { sessionId: live.sessionId, api: live.api, logger: live.logger }
-        : live,
-    )
-    await emitReflectionHealthAlert(completionsDir, this.options.identity.id, live, this.warnedHealth)
-    return {
-      runId: run.runId,
-      outcome: result.outcome,
-      ...(result.reason === undefined ? {} : { reason: result.reason }),
-      ...(result.detail === undefined ? {} : { detail: result.detail }),
-      completion,
-      ...(transition.launch === undefined ? {} : { launch: transition.launch }),
-    }
+      resolution,
+      suppressCompletionNotification,
+      options: this.options,
+      now: this.now,
+      ensureRenderer: (live) => this.ensureRenderer(live),
+      warnedHealth: this.warnedHealth,
+    })
   }
 
-  private async deliverFinalized(result: ReflectionRunResult): Promise<ReflectionRunResult> {
-    const live = this.options.liveSession?.()
-    this.ensureRenderer(live)
-    const finishedAt = result.completion.finishedAt
-    const health = await readReflectionHealth(join(this.options.identity.paths.reflection, "completions"))
-    const completionsDir = join(this.options.identity.paths.reflection, "completions")
-    const completion = await recordReflectionCompletion(
-      completionsDir,
-      {
-        ...result.completion,
-        durationMs: Math.max(0, Date.parse(finishedAt) - Date.parse(result.completion.startedAt)),
-        ...(result.completion.outcome === "merged"
-          ? await this.mergedMetadata(result.completion.runId)
-          : {}),
-        consecutiveFailures: result.completion.outcome === "failed" ? health.streak : 0,
-      },
-      live,
-    )
-    await emitReflectionHealthAlert(completionsDir, this.options.identity.id, live, this.warnedHealth)
-    return { ...result, completion }
+  private deliverFinalized(result: ReflectionRunResult): Promise<ReflectionRunResult> {
+    return publishFinalizedReflectionRun({
+      result,
+      options: this.options,
+      ensureRenderer: (live) => this.ensureRenderer(live),
+      mergedMetadata: (runId) => this.mergedMetadata(runId),
+      warnedHealth: this.warnedHealth,
+    })
   }
 
   private async appendLaunched(
@@ -317,11 +167,6 @@ export class SenpiSubprocessRunner implements ReflectionRunner {
     if (!live || this.registeredApis.has(live.api)) return
     registerReflectionCompletionRenderer(live.api)
     this.registeredApis.add(live.api)
-  }
-
-  private async discard(worktree: ReflectionWorktree): Promise<ReflectionFinalizeResult> {
-    const cleanup = await cleanupReflectionWorktree(worktree)
-    return { status: "failed", cleanup }
   }
 
   private finalizationContext(): RunFinalizationContext {
